Unit tests creation and trigger update

+++ force-app/main/default/triggers/BookTrigger.trigger
@@ -1,3 +1,7 @@
 trigger BookTrigger on Book__c (before delete) {
-    TriggerHandler.deleteBooks(Trigger.old);
+    if(Trigger.isBefore){
+        if(Trigger.isDelete){
+            TriggerHandler.deleteBooks(Trigger.old);
+        }
+    }
 }
